move graph_capture into parallel state
